Owned Shoe is currently being worked on for the bridge pattern
