+++ Abstraction/Owner/ControllerModelOwnerWithArray.cs
@@ -6,63 +6,71 @@ using FastTrackEServices.HelperAlgorithms;
 using FastTrackEServices.Implementation;
 using Microsoft.AspNetCore.Mvc;
 
-public class ControllerModelOwnerWithArray : ControllerBase {
-    IGet get;
-    IPut put;
-    IDelete delete;
-    IPost post;
-    ITransform transform;
+abstract public class ControllerModelOwnerWithArray : ControllerBase {
+
     private readonly AppDbContext context;
 
-    public ControllerModelOwnerWithArray (ITransform transform, IGet get, IPost post, IPut put, IDelete delete, AppDbContext context)
+    protected readonly IEnumerable<IRestOperation> services;
+
+    protected readonly IEnumerable<IRestOperation> restOperation;
+    
+    public ControllerModelOwnerWithArray (AppDbContext context, IEnumerable<IRestOperation> services)
     {
         this.context = context;
-        this.get = get;
-        this.post = post;
-        this.put = put;
-        this.delete = delete;
-        this.transform = transform;
+        this.services = services;
     }
 
-    async public Task<IActionResult>  Post(Object dto, string entityName, string entityType)
-    {
-        Dictionary<string, object> result = await this.post.post(context, dto);
+    abstract public Task<IActionResult>  Post(Object dto);
 
-        if (result["Result"] != "Success")
-        return StatusCode(400, new {data = result});
+    abstract public Task<IActionResult> GetAll();
 
-        return StatusCode(201, new {data = $"{entityType} {entityName} has been successfully created!"});
-    }
-    async public Task<IActionResult> GetAll()
-    {
-        Console.WriteLine("Request Path: " + Request.Path);
-        Console.WriteLine("Function called: " + this.get.GetType());
-        Dictionary<string, object> result = await this.get.GetAll(this.context);
-        return StatusCode(200, new { data = result });
-    }
+    abstract public Task<IActionResult> Get(int id, string entityType);
 
-    async public Task<IActionResult> Get(int id, string entityType)
-    {
-        Object result = await this.get.Get(this.context, id);
-        if (result == null)
-        {
-            return StatusCode(200, new { data = $"{entityType} with an ID of {id} is not found" });
-        }
-        return StatusCode(200, new {data = result});
-    }
+    abstract public Task<IActionResult> Put(Object dto, ITransform transform);
 
-    async public Task<IActionResult> Put(Object dto, ITransform transform)
-    {
-        Dictionary<string, object> result = await this.put.put(this.context, dto, transform);
-        if (result["Result"] != "Success")
-        return StatusCode(400, new {data = result});
+    abstract public Task<IActionResult> Delete(int id);
 
-        return StatusCode(200, new {data = result});
-    }
 
-    async public Task<IActionResult> Delete(int id)
-    {
-        await this.delete.delete(this.context, id);
-        return StatusCode(204);
-    }
+    // async public Task<IActionResult>  Post(Object dto, string entityName, string entityType)
+    // {
+    //     Dictionary<string, object> result = await this.restOperation.post();
+
+    //     if (result["Result"] != "Success")
+    //     return StatusCode(400, new {data = result});
+
+    //     return StatusCode(201, new {data = $"{entityType} {entityName} has been successfully created!"});
+    // }
+
+    // async public Task<IActionResult> GetAll()
+    // {
+    //     Dictionary<string, object> result = await this.get.GetAll(this.context);
+    //     Console.WriteLine("Request Path: " + Request.Path);
+    //     Console.WriteLine("Function called: " + this.get.GetType());
+    //     return StatusCode(200, new { data = result });
+    // }
+
+    // async public Task<IActionResult> Get(int id, string entityType)
+    // {
+    //     Object result = await this.get.Get(this.context, id);
+    //     if (result == null)
+    //     {
+    //         return StatusCode(200, new { data = $"{entityType} with an ID of {id} is not found" });
+    //     }
+    //     return StatusCode(200, new {data = result});
+    // }
+
+    // async public Task<IActionResult> Put(Object dto, ITransform transform)
+    // {
+    //     Dictionary<string, object> result = await this.put.put(this.context, dto, transform);
+    //     if (result["Result"] != "Success")
+    //     return StatusCode(400, new {data = result});
+
+    //     return StatusCode(200, new {data = result});
+    // }
+
+    // async public Task<IActionResult> Delete(int id)
+    // {
+    //     await this.delete.delete(this.context, id);
+    //     return StatusCode(204);
+    // }
 }
