+++ Controller/Shoe/ShoeController.cs
@@ -1,7 +1,6 @@
-namespace Controller;
+// namespace Controller;
 using FastTrackEServices.Data;
 using FastTrackEServices.DTO;
-using FastTrackEServices.Model;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using FastTrackEServices.Abstraction;
@@ -9,55 +8,60 @@ using FastTrackEServices.Implementation;
 using Implementation.Concrete;
 using System.Text.Json;
 using FastTrackEServices.HelperAlgorithms;
-using System.Reflection;
 
 [Route("api/[controller]")]
 [ApiController]
-public class ShoeController : ControllerModelOwnerWithArray {
+public class ShoeController : ControllerModelOwner {
     private const string constantIndividualPath = "[action]/{id:int}";
 
-    private readonly AppDbContext appDbContext;
+    private readonly AppDbContext context;
 
-    private readonly IGet get;
-    
-    private readonly IPut put;
+    protected readonly IEnumerable<IRestOperation> services;
 
-    private readonly IDelete delete;
+    protected readonly IRestOperation restOperation;
 
-    private readonly IPost post;
-
-    private readonly ITransform transform;
-
-    public ShoeController(ITransform transform, IGet get, IPost post, IPut put, IDelete delete, AppDbContext context) : base (transform, get, post, put, delete, context)
+    public ShoeController(AppDbContext context, IEnumerable<IRestOperation> services) : base (context, services)
     {
-        appDbContext = context;
-        this.transform = new CollectionToStringArray();
-        this.get = new ShoeGet();
-        this.post = new ShoePost();
-        this.delete = new ShoeDelete();
-        this.put = new ShoePut();
+        this.context = context;
+        this.services = services;
+        this.restOperation = services.FirstOrDefault(s => s.GetType() == typeof(ShoeRest));
     }
 
     [HttpGet("[action]")]
-    public Task<IActionResult> GetAll()
+    async public override Task<IActionResult> GetAll()
     {
-        return base.GetAll();
+
+        Dictionary<string, object> result = await this.restOperation.GetAll(this.context);
+        return StatusCode(200, new { data = result });
     }
 
     [HttpGet($"{constantIndividualPath}")]
-    public Task<IActionResult> Get([FromRoute] int id)
+    async public override Task<IActionResult> Get([FromRoute] int id)
     {  
-        return base.Get(id, this.ControllerContext.RouteData.Values["controller"].ToString());
+        Object result = await this.restOperation.Get(this.context, id); 
+        string entityType = this.ControllerContext.RouteData.Values["controller"].ToString();
+        if (result == null)
+        {
+            return StatusCode(200, new { data = $"{entityType} with an ID of {id} is not found" });
+        }
+        return StatusCode(200, new {data = result});
     }
 
     [HttpPost("[action]")]
-    async public Task<IActionResult> Post([FromBody] Object dto)
+    async public override Task<IActionResult> Post([FromBody] Object dto)
     {
         try {
-            string shoeType = this.ControllerContext.RouteData.Values["controller"].ToString();
-            string shoeName = JsonSerializer.Deserialize<CreateShoe>(dto.ToString()).name;
-            Task<IActionResult> result = base.Post(dto, shoeName, shoeType);
-            return result.Result;
+            string clientType = this.ControllerContext.RouteData.Values["controller"].ToString();
+            string clientName = JsonSerializer.Deserialize<CreateClient>(dto.ToString()).username;
+
+            Dictionary<String, Object> result = await this.restOperation.Post(this.context, dto);
+
+            if (result["Result"] == "Success")
+            return StatusCode(201, new {data = $"{clientType} has been successfully created!"});
+
+            return StatusCode(400, new {data = result});
+
+
         } 
         
         catch (DbUpdateException ex)
@@ -67,13 +71,17 @@ public class ShoeController : ControllerModelOwnerWithArray {
     }
 
     [HttpPut("[action]")]
-    async public Task<IActionResult> Put([FromBody] Object dto)
+    async public override Task<IActionResult> Put([FromBody] Object dto, ITransform transform)
     {
         // Assuming that the DTO's shoeColor array contains a hashmap of <id - int, string - name>
         // Frontend returns the original id for a shoecolor, but possibly with a different name
         try {
-            Task<IActionResult> result = base.Put(dto, transform);
-            return result.Result;
+            Dictionary<String, Object> result = this.restOperation.Put(this.context, dto, transform).Result;
+            
+            if (result["Result"] != "Success")
+            return StatusCode(400, new {data = result});
+
+            return StatusCode(200, new {data = result});
         }
 
         catch (DbUpdateException ex)
@@ -84,10 +92,11 @@ public class ShoeController : ControllerModelOwnerWithArray {
     }
 
     [HttpDelete($"{constantIndividualPath}")]
-    async public Task<IActionResult> Delete([FromRoute] int id)
+    async public override Task<IActionResult> Delete([FromRoute] int id)
     {
         try {
-             return await base.Delete(id);
+             await this.restOperation.Delete(this.context, id);
+             return StatusCode(204);
         } catch (DbUpdateException ex)
         {
             return StatusCode(500, new {data = ex.InnerException.Message});
