+++ Implementation/Concrete/OwnedShoe/OwnedShoeRest.cs
@@ -1,0 +1,13 @@
+using System.Text.Json;
+using FastTrackEServices.Data;
+using FastTrackEServices.DTO;
+using FastTrackEServices.HelperAlgorithms;
+using FastTrackEServices.Model;
+using Microsoft.EntityFrameworkCore;
+
+namespace FastTrackEServices.Implementation;
+
+public class OwnedShoeRest : IRestOperation {
+
+    
+}
