+++ bin/Debug/net7.0/appsettings.json
@@ -8,5 +8,12 @@
   "AllowedHosts": "*",
   "ConnectionStrings": {
     "Default" : "server=localhost;database=SoftwareEngineering;User=root;Password=;"
-  }
+  },
+
+  
+    "LabelGenOptions": {
+      "Prefix": "CD",
+      "Suffix": "MZ"
+    }
+  
 }
